Relocate real function sandbox journeys to the function shard
